+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from '@/structures/linked-list'
 
 export * from '@/structures/stack'
+
+export * from '@/structures/queue'