+++ src/structures/queue.ts
@@ -2,15 +2,16 @@ import { ListNode } from '@/core/list-node'
 
 type Constructor<T, A extends unknown[] = unknown[]> = new (...args: A) => T
 
-export class Stack<T, A extends unknown[] = unknown[]> {
+export class Queue<T, A extends unknown[] = unknown[]> {
   #head: ListNode<T> | null
+  #tail: ListNode<T> | null
   public length: number
   #TypeCtor?: Constructor<T, A>
   constructor(values?: T[], TypeCtor?: Constructor<T, A>) {
-    this.#head = null
+    this.#head = this.#tail = null
     this.length = 0
-    this.#initStack(values)
     this.#TypeCtor = TypeCtor
+    this.#initQueue(values)
   }
 
   // Generator function
@@ -24,24 +25,26 @@ export class Stack<T, A extends unknown[] = unknown[]> {
 
   public push(val: T): void {
     const newNode = new ListNode(val)
-    this.length++
 
     // Empty Stack
-    if (this.#head == null) {
-      this.#head = newNode
+    if (this.isEmpty()) {
+      this.#head = this.#tail = newNode
+      this.length++
       return
     }
     // Not Empty
-    newNode.next = this.#head
-    this.#head = newNode
+    this.#tail!.next = newNode
+    this.#tail = newNode
+    this.length++
   }
 
   public pop(): T | undefined {
-    if (this.#head == null) return
+    if (this.isEmpty()) return
     this.length--
-    const popped = this.#head
+    const popped = this.#head!
     const val = popped.val
-    this.#head = this.#head.next
+    this.#head = this.#head!.next
+    if (this.#head === null) this.#tail = null
     popped.cleanup()
     return val
   }
@@ -53,11 +56,11 @@ export class Stack<T, A extends unknown[] = unknown[]> {
   }
 
   public static equals<U, V extends unknown[] = unknown[]>(
-    stack1: Stack<U, V>,
-    stack2: Stack<U, V>,
+    queue1: Queue<U, V>,
+    queue2: Queue<U, V>,
     comparator: (a: U, b: U) => boolean = (a, b) => a === b,
   ): boolean {
-    if (stack1.length !== stack2.length) return false
+    if (queue1.length !== queue2.length) return false
     function backtrack(
       head1: ListNode<U> | null,
       head2: ListNode<U> | null,
@@ -68,34 +71,47 @@ export class Stack<T, A extends unknown[] = unknown[]> {
       if (!comparator(head1.val, head2.val)) return false
       return backtrack(head1.next, head2.next)
     }
-    return backtrack(stack1.#head, stack2.#head)
+    return backtrack(queue1.#head, queue2.#head)
   }
 
   public static swap<U, V extends unknown[] = unknown[]>(
-    stack1: Stack<U, V>,
-    stack2: Stack<U, V>,
+    queue1: Queue<U, V>,
+    queue2: Queue<U, V>,
   ): void {
-    if (!(stack1 instanceof Stack && stack2 instanceof Stack)) {
-      throw new TypeError('Input value must be a stack data structure')
+    if (!(queue1 instanceof Queue && queue2 instanceof Queue)) {
+      throw new TypeError('Input value must be a queue data structure')
     }
-    const tempHead = stack1.#head
-    const tempLength = stack1.length
+    const tempHead = queue1.#head
+    const tempTail = queue1.#tail
+    const tempLength = queue1.length
 
-    stack1.#head = stack2.#head
-    stack1.length = stack2.length
+    queue1.#head = queue2.#head
+    queue1.#tail = queue2.#tail
+    queue1.length = queue2.length
 
-    stack2.#head = tempHead
-    stack2.length = tempLength
+    queue2.#head = tempHead
+    queue2.#tail = tempTail
+    queue2.length = tempLength
   }
 
-  public get top(): T | undefined {
-    if (this.#head == null) return
-    return this.#head.val
+  public get front(): T | undefined {
+    if (this.isEmpty()) return
+    return this.#head!.val
   }
 
-  public set top(val: T) {
-    if (this.#head == null) return
-    this.#head.val = val
+  public set front(val: T) {
+    if (this.isEmpty()) return
+    this.#head!.val = val
+  }
+
+  public get back(): T | undefined {
+    if (this.isEmpty()) return
+    return this.#tail!.val
+  }
+
+  public set back(val: T) {
+    if (this.isEmpty()) return
+    this.#tail!.val = val
   }
 
   public isEmpty(): boolean {
@@ -109,7 +125,7 @@ export class Stack<T, A extends unknown[] = unknown[]> {
       current.cleanup()
       current = next
     }
-    this.#head = null
+    this.#head = this.#tail = null
     this.length = 0
   }
 
@@ -117,10 +133,10 @@ export class Stack<T, A extends unknown[] = unknown[]> {
     return [...this]
   }
 
-  #initStack(values?: T[]) {
+  #initQueue(values?: T[]) {
     if (!values) return
     if (!Array.isArray(values)) {
-      throw new TypeError('Invalid argument type: new Stack()')
+      throw new TypeError('Invalid argument type: new Queue()')
     }
     if (values.length === 0) return
     for (const value of values) {