+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
 
+## [3.0.0] - 2025-06-30
+
+### Added
+
+- Data Structure: `Queue`
+- Documentation: `docs/queue.md` for detailed API docs
+- Added new method .toArray() in Stack and Queue
+
+---
+
 ## [2.0.0] - 2025-06-30
 
 ### Added